% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomadlist.R
\name{getnomadcities}
\alias{getnomadcities}
\title{Returns a dataframe based on the list cities request.
#'}
\usage{
getnomadcities()
}
\value{
a dataframe with all nomad cities.
}
\description{
Returns a dataframe based on the list cities request.
#'
}
\examples{
getnomadcities()

}

